refactor: update bolt-events import sites to @bigbluebam/shared
Rewires every consuming file to import publishBoltEvent from @bigbluebam/shared
instead of the deleted per-service ../lib/bolt-events.js (or
../utils/bolt-events.js for the worker). Apps that did not already declare a
workspace dependency on @bigbluebam/shared get one added (blast-api keeps its
local buildCampaignEventPayload import via the renamed enrich-campaign-event.ts;
board-api keeps buildBoardEventPayload via enrich-board-event.ts).

Updated import sites by app:

  apps/api: services/task.service.ts, routes/sprint.routes.ts, routes/comment.routes.ts
  apps/banter-api: routes/channel.routes.ts, routes/message.routes.ts,
                   routes/reaction.routes.ts, routes/thread.routes.ts
  apps/beacon-api: routes/beacon.routes.ts
  apps/bearing-api: routes/goals.ts, routes/key-results.ts
  apps/bill-api: routes/invoices.routes.ts, routes/payments.routes.ts,
                 services/payment.service.ts
  apps/blank-api: routes/forms.routes.ts, routes/public.routes.ts
  apps/blast-api: routes/campaigns.routes.ts (split: shared publishBoltEvent +
                  local enrich-campaign-event)
  apps/board-api: routes/board.routes.ts (split: shared publishBoltEvent +
                  local enrich-board-event)
  apps/book-api: routes/events.routes.ts, routes/public-booking.routes.ts
  apps/brief-api: services/document.service.ts

Also annotates apps/bolt-api/src/lib/publish-event.ts with a BOLT-API
INTERNAL ONLY header documenting why it stays separate from the shared
HTTP publisher: it is a Redis PubSub fan-out variant used only inside
bolt-api, with a different transport and a different argument shape.

Adds @bigbluebam/shared as a workspace:* dep in 9 apps (banter-api, beacon-api,
bill-api, blank-api, blast-api, board-api, bond-api, book-api, brief-api) that
did not previously depend on it; pnpm-lock.yaml is regenerated to match.

Bond signature changes and the worker signature change ship in the next
commit so the diff stays readable.

Part of Wave 0.3 / P0.2 from Cross_Product_Integration_Plan.md Step 3.
