+++ apps/blank-api/src/routes/forms.routes.ts
@@ -2,7 +2,7 @@ import type { FastifyInstance } from 'fastify';
 import { z } from 'zod';
 import { requireAuth, requireMinRole, requireScope } from '../plugins/auth.js';
 import * as formService from '../services/form.service.js';
-import { publishBoltEvent } from '../lib/bolt-events.js';
+import { publishBoltEvent } from '@bigbluebam/shared';
 import { enrichForm, loadActor, loadOrg } from '../lib/bolt-enrich.js';
 
 // ---------------------------------------------------------------------------
