+++ apps/book-api/package.json
@@ -12,6 +12,7 @@
     "lint": "biome check ./src"
   },
   "dependencies": {
+    "@bigbluebam/shared": "workspace:*",
     "argon2": "^0.41.1",
     "fastify": "^5.2.0",
     "@fastify/cors": "^10.0.0",
