+++ apps/book-api/src/routes/events.routes.ts
@@ -2,7 +2,7 @@ import type { FastifyInstance } from 'fastify';
 import { z } from 'zod';
 import { requireAuth, requireMinRole, requireScope } from '../plugins/auth.js';
 import * as eventService from '../services/event.service.js';
-import { publishBoltEvent } from '../lib/bolt-events.js';
+import { publishBoltEvent } from '@bigbluebam/shared';
 import { enrichEvent, loadActor, loadOrg } from '../lib/bolt-enrich.js';
 
 const createEventSchema = z.object({
